Repackage animated navigation classes, add more kdoc with examples
