+++ circuitx/gesture-navigation/src/browserCommonMain/kotlin/com/slack/circuitx/gesturenavigation/GestureNavigationDecoration.js.kt
@@ -2,7 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 package com.slack.circuitx.gesturenavigation
 
-import com.slack.circuit.foundation.AnimatedNavDecorator
+import com.slack.circuit.foundation.animation.AnimatedNavDecorator
 
 public actual fun GestureNavigationDecorationFactory(
   fallback: AnimatedNavDecorator.Factory,
